Fix escape character
`If you haven\'t done so already` => `If you haven't done so already`

It appeared verbatim as `haven\'t done` on my phone, so I guess the escape character has to be removed? Not sure if this is the right fix. Feel free to ignore this pr if the correct fix is elsewhere.